Refactoring of file structure
