+++ lib/Notes/NoteCreatePage/NoteCreatePage.js
@@ -10,9 +10,9 @@ import {
 
 import { Icon } from '@folio/stripes-components';
 
-import NoteForm from '../NoteForm';
+import NoteForm from '../components/NoteForm';
 import { noteTypesCollectionShape } from '../response-shapes';
-import { referredEntityDataShape } from '../NoteForm/noteShapes';
+import { referredEntityDataShape } from '../components/NoteForm/noteShapes';
 
 @stripesConnect
 class NoteCreatePage extends Component {
