--- lib/Notes/NoteForm/index.js
@@ -1,2 +1,0 @@
-export { default } from './NoteForm';
-export { default as NoteView } from './NoteView';
