+++ lib/Notes/NoteViewPage/NoteViewPage.js
@@ -12,10 +12,10 @@ import {
   ConfirmationModal,
 } from '@folio/stripes-components';
 
-import AssignmentsList from '../AssignmentsList';
-import { NoteView } from '../NoteForm';
+import AssignmentsList from '../components/AssignmentsList';
+import NoteView from './components/NoteView';
 import { noteShape } from '../response-shapes';
-import { referredEntityDataShape } from '../NoteForm/noteShapes';
+import { referredEntityDataShape } from '../components/NoteForm/noteShapes';
 
 @stripesConnect
 class NoteViewPage extends Component {
