+++ lib/Notes/NoteViewPage/components/NoteView/NoteView.css
@@ -1,0 +1,4 @@
+.note-view-content {
+  margin: 0 auto;
+  max-width: 50rem;
+}
