+++ lib/Notes/NoteViewPage/components/NoteView/NoteView.js
@@ -19,15 +19,15 @@ import {
   KeyValue,
 } from '@folio/stripes-components';
 
-import { NOTE_LINKS_MIN_NUMBER } from '../constants';
-import AssignmentsList from '../AssignmentsList';
-import ReferredRecord from '../ReferredRecord';
-import styles from './NoteForm.css';
+import { NOTE_LINKS_MIN_NUMBER } from '../../../constants';
+import AssignmentsList from '../../../components/AssignmentsList';
+import ReferredRecord from '../../../components/ReferredRecord';
+import styles from './NoteView.css';
 
 import {
   noteDataShape,
   referredEntityDataShape,
-} from './noteShapes';
+} from '../../../components/NoteForm/noteShapes';
 
 export default class NoteView extends Component {
   static propTypes = {
@@ -214,7 +214,7 @@ export default class NoteView extends Component {
             <Pane
               lastMenu={this.renderExpandAllButton()}
               defaultWidth="50%"
-              className={styles['note-form-content']}
+              className={styles['note-view-content']}
             >
               <Accordion
                 label={<FormattedMessage id="stripes-smart-components.notes.generalInformation" />}
