+++ lib/Notes/NoteViewPage/components/NoteView/index.js
@@ -1,0 +1,1 @@
+export { default } from './NoteView';
