+++ lib/Notes/NotesSmartAccordion/NotesSmartAccordion.js
@@ -14,7 +14,7 @@ import {
   sortOrders,
   notesStatuses,
 } from '../constants';
-import NotesAccordion from '../NotesAccordion';
+import NotesAccordion from '../components/NotesAccordion';
 
 import { makeSearchQuery } from './utils';
