+++ lib/Notes/components/AssignmentsList/AssignmentsList.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
 
 import { List } from '@folio/stripes-components';
-import getLinkedEntityTypesArr from '../utils/getLinkedEntityTypesArr';
+import getLinkedEntityTypesArr from '../../utils/getLinkedEntityTypesArr';
 
 import styles from './AssignmentsList.css';
