+++ lib/Notes/components/NavigationModal/index.js
