+++ lib/Notes/components/NoteForm/NoteForm.js
@@ -32,7 +32,7 @@ import {
 import {
   MAX_TITLE_LENGTH,
   MAX_DETAILS_LENGTH,
-} from '../constants';
+} from '../../constants';
 
 import styles from './NoteForm.css';
