+++ lib/Notes/components/NoteForm/index.js
@@ -1,0 +1,2 @@
+export { default } from './NoteForm';
+
