+++ lib/Notes/components/NotesAccordion/components/NotesAssigningModal/NotesAssigningModal.js
@@ -22,13 +22,13 @@ import {
 import {
   SearchAndSortResetButton as ResetButton,
   CheckboxFilter,
-} from '../../../../..';
+} from '../../../../../..';
 
 import {
   notesStatuses,
   sortOrders,
   NOTE_LINKS_MIN_NUMBER,
-} from '../../../constants';
+} from '../../../../constants';
 
 import {
   columnNames,
